足回り左右モードの追加、いとこを操縦者に優しく変更

+++ src/diff_angle_calc.hpp
@@ -2,9 +2,9 @@
 #define DIFF_ANGLE_CALC_HPP
 
 #include "mbed.h"
-int16_t diff_angle_calc(int16_t current_angle, int16_t previous_angle)
+int32_t diff_angle_calc(int16_t current_angle, int16_t previous_angle)
 {
-    int16_t diff = current_angle - previous_angle;
+    int32_t diff = current_angle - previous_angle;
     if (diff > 4096)
     {
         diff -= 8192;
